[FEATURE] Adding a translation
The analysis of menu generation was a good and important step as a basis for performance optimizations - also in the core itself. So far, however, the issue of translation has been ignored.

A complete translation of the demo website has now been carried out.
After importing the SQL dump, around 1000 pages per language are now available. Furthermore, a language switcher has been added to the layout.